requirements.txt
flask
Pillow
transformers
torch
